Move scalacheck integration into the toplevel package
