+++ munit-scalacheck/shared/src/main/scala/munit/ScalaCheckFailException.scala
@@ -1,4 +1,4 @@
-package munit.scalacheck
+package munit
 
 import scala.util.control.NoStackTrace
